packages/anumate-oidc/pyproject.toml
[tool.poetry]
name = "anumate-oidc"
version = "0.1.0"
description = "OIDC token verifier for the Anumate platform."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
PyJWT = "^2.4.0"
requests = "^2.28.0"
fastapi = "^0.104.0"
anumate-crypto = {path = "../anumate-crypto", develop = true}

[tool.poetry.dev-dependencies]
pytest = "^7.0"
respx = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
